requirements.txt
asyncpg==0.29.0
alembic==1.11.1
bcrypt==4.0.1
cryptography==41.0.3
celery==5.3.1
databases==0.9.0
fastapi==0.100.0
fastapi-limiter==0.1.6
gunicorn==21.2.0
itsdangerous==2.1.2
passlib==1.7.4
pandas==2.2.0
python-jose==3.3.0
pydantic==2.0.3
pydantic_core==2.3.0
python-decouple==3.8
pysftp==0.2.9
redis==4.6.0
SQLAlchemy==2.0.19
sqladmin==0.14.0
uvicorn==0.23.0
psycopg2-binary==2.9.9
requests==2.31.0
Jinja2==3.1.3
urllib3==1.26.6
fastapi-mail==1.4.1
openpyxl==3.1.2
XlsxWriter==3.2.0
pycryptodome==3.20.0

